Update lab 1
